backend/requirements.txt
# Astitva AI Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML & Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
timm>=0.9.0
numpy>=1.24.0
scipy>=1.10.0

# Optional - for additional image processing
# opencv-python>=4.8.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

